eliminated the duplicates
